+++ src/Network/Misskey/Api/Users/Notes.hs
@@ -18,12 +18,10 @@ module Network.Misskey.Api.Users.Notes
 import Lens.Simple (makeLenses, (^.))
 import Data.Time (UTCTime)
 import Data.Aeson ((.=), object)
-import System.Posix.Types (EpochTime)
-import Data.Time.Clock.POSIX (utcTimeToPOSIXSeconds)
-import Foreign.C.Types (CTime(..))
+
 
 import Network.Misskey.Type
-import Network.Misskey.Api.Internal (postRequest, createObj, createMaybeObj)
+import Network.Misskey.Api.Internal (postRequest, createObj, createMaybeObj, createUTCTimeObj)
 
 data APIRequest = APIRequest { _userId           :: Id
                              , _includeReplies   :: Bool
@@ -42,24 +40,20 @@ makeLenses ''APIRequest
 usersNotes :: APIRequest -> Misskey [Note]
 usersNotes req = postRequest "/api/users/notes" body
     where
-        userIdObj           = createObj      "userId"           (req^.userId)
-        includeRepObj       = createObj      "includeReplies"   (req^.includeReplies)
-        limitObj            = createMaybeObj "limit"            (req^.limit)
-        sinceIdObj          = createMaybeObj "sinceId"          (req^.sinceId)
-        untilIdObj          = createMaybeObj "untilId"          (req^.untilId)
-        sinceDateObj        = maybe [] (\x -> ["sinceDate" .= uToE x]) (req^.sinceDate)
-        untilDateObj        = maybe [] (\x -> ["untilDate" .= uToE x]) (req^.untilDate)
-        includeMyRenotesObj = createObj      "includeMyRenotes" (req^.includeMyRenotes)
-        withFilesObj        = createObj      "withFiles"        (req^.withFiles)
-        fileTypeObj         = createMaybeObj "fileType"         (req^.fileType)
-        excludeNsfwObj      = createObj      "excludeNsfw"      (req^.excludeNsfw)
+        userIdObj           = createObj        "userId"           (req^.userId)
+        includeRepObj       = createObj        "includeReplies"   (req^.includeReplies)
+        limitObj            = createMaybeObj   "limit"            (req^.limit)
+        sinceIdObj          = createMaybeObj   "sinceId"          (req^.sinceId)
+        untilIdObj          = createMaybeObj   "untilId"          (req^.untilId)
+        sinceDateObj        = createUTCTimeObj "sinceDate"        (req^.sinceDate)
+        untilDateObj        = createUTCTimeObj "untilDate"        (req^.untilDate)
+        includeMyRenotesObj = createObj        "includeMyRenotes" (req^.includeMyRenotes)
+        withFilesObj        = createObj        "withFiles"        (req^.withFiles)
+        fileTypeObj         = createMaybeObj   "fileType"         (req^.fileType)
+        excludeNsfwObj      = createObj        "excludeNsfw"      (req^.excludeNsfw)
         body                = mconcat [userIdObj, includeRepObj, limitObj, sinceIdObj
                                       , untilIdObj, sinceDateObj, untilDateObj, includeMyRenotesObj
                                       , withFilesObj, fileTypeObj, excludeNsfwObj]
 
 
--- | Convert UTCTime to UNIX time
--- 
--- This code is from: https://kazu-yamamoto.hatenablog.jp/entry/20130329/1364525770
-uToE :: UTCTime -> EpochTime
-uToE = CTime . truncate . utcTimeToPOSIXSeconds
+